README: create public user in Magpie for JupyterHub login
For the public demo account login on JupyterHub.

birdhouse/config/magpie/permissions.cfg will only create the user if it
needs to be created for a specific permission on a resource or service.

The config doesn't handle defining only a user/group by itself without
them, see
https://github.com/Ouranosinc/Magpie/blob/9f31c9543f00048441675705cfb3432c63ba6991/config/permissions.cfg#L1-L26

https://github.com/Ouranosinc/Magpie/blob/9f31c9543f00048441675705cfb3432c63ba6991/magpie/helpers/create_users.py

The create_users.py script above hardcode the username same as the user
email and a random password, not what we need.

We'll just add a manual steps to the README for now.