proxy: add mechanism to override canarie-api monitoring config
Drop a .py file into CANARIE_MONITORING_EXTRA_CONF_DIR inside the proxy
container to be able to override anything in docker_configuration.py.template.

The order of loading the extra config files are alphabetically sorted so
the last one have highest override precedence.

The logging are not shown anywhere right now but are available if
needed.