scheduler: dummy minimal implementation to understand and assess its capabilities
Re-use an existing crontab in docker that can spawn new container
(tested here) or even use existing container (not tested yet) to do the
work.  Pretty nice.

All useful logs are in `docker logs`, awsome, one centralized place.

It is implemented as a separate components because maybe, initially, not
everyone will want to enable what the scheduler will do (autodeploy).

Also, long term, we want to allow cherry-picking the multiple components
of PAVICS so not everyone is forced to deploy the entire PAVICS stack so
a separate component make sense.