autodeploy: fix deploy key dir not working inside container
Broken when we tried to ensure the autodeploy is backward-compat when
not running inside the new "scheduler" container.

The autodeploy code will re-load the common.env and env.local file
inside the "scheduler" container so the list of environment variable
available and their value is different inside and outside of the
"scheduler" container.

By performing template expansion AUTODEPLOY_DEPLOY_KEY_ROOT_DIR we
"freeze" the good value outside of the "scheduler" container, then we
can restore the good value after the re-load.

When not running inside the new "scheduler" container, the re-load will
always get the good "outside" value so no problem here so not need to
repeat default values again inside deploy.sh and triggerdeploy.sh.