testthredds: add new TestWps_Outputs dataset catalog
For Emu test bird to write to (require manual change to Emu config).

The old wps_outputs mount was juste to make it available inside
testthredds container without exposing it as a catalog.

All Thredds catalog volume mount are read-only because Thredds do not
have write capabilities to those catalogs.