finch2: add Postgres DB for proper WPS logging of Async WPS requests
Default was SQLite in-memory and it does not clear completed async wps
requests so we quickly run into the default 2 max concurrent wps
requests !

Use a separate DB so this optional component do not leave traces when we
disable it.

This will also allow us to test new different version of Postgres DB if
needed.  Default version is latest Postgres, not the old version
currently in used.