finch2: make this component generic to accommodate any other WPS than Finch
We can plug in other WPS docker images in this skeleton/generic
component.  All the elements that need to change are configurable (was
just missing the FINCH2_INTERNAL_PORT).

The HOSTNAME and HTTP_PORT env var are only for older birds so disabled
them.

Set WPS logging to a more suitable production logging level, can easily
be changed later.