monitoring: fix system load grafana dashboard
Was not showing any data before because variable $server was not
defined correctly.

The previous query using node_cpu do not exist anymore so had to
simplify the query to not need it anymore.

Now graph show raw load values, not converting to percent depending on
number of cpus anymore.  Didn't know how to get the number of cpus.