jupyterhub: update to version 1.3.0 to include login terms patch
This version of jupyterhub includes the login terms patch originally
introduced in 8be8eeac211d3f5c2de620781db8832fdb8f9093.

This official login terms feature has a few enhancements:
* no javascript dependency
* pop-up reminder for user to check the checkbox

Behavior change is the "Sign in" button is not longer disabled if
unchecked.  It simply does not work and reminds the user to check the
checkbox if unchecked.