geosever: fix data dir permission was always running when under autodeploy
Autodeploy runs inside its own container so `/data/geoserver` dir never
exists in its container, which always trigger the
`fix-geoserver-data-dir-perm` script.

The fix is to check whether a file under `/data/geoserver` exists
instead of the dir itself.

Preserved the existing default mode of operation of the script to always
perform the permmission fix for the case when upgrading with existing
data.  This case of course requires a manual run of the script.

When the script is invoked automatically by `pavics-compose.sh`, it will
only run on first startup of Geoserver, whether `pavics-compose.sh` is
called manually by hand or automatically from inside autodeploy container.