jupyterhub: allow config override via env.local
This is basically the same as `ENABLE_JUPYTERHUB_MULTI_NOTEBOOKS` but at
the bottom of the file so it can override everything.
`ENABLE_JUPYTERHUB_MULTI_NOTEBOOKS` is kept for backward-compat.

First useful application is to enable server culling for auto shutdown
of idle kernels and jupyter single server (fixes https://github.com/bird-house/birdhouse-deploy/issues/67).