jupyter: fix README file always volume mount even when do not exist on disk (BC)
The faulty behavior is due to the JUPYTERHUB_README volume-mount in the
`jupyterhub` container.  If the JUPYTERHUB_README file do not exist on
disk, it will always volume-mount, resulting in creating an empty dir
with name JUPYTERHUB_README.

Basically the previous check whether the JUPYTERHUB_README exist on disk
before volume-mount in the user Jupyter server was useless since it will
always exist due to the forced volume-mount.

The previous behavior also assume there is always a JUPYTERHUB_README
file which will not be true in this PR since the JUPYTERHUB_README in
this repo is fairly specific to Ouranos and will be moved out.

This change is not backward-compatible since the default value has
changed.

To restore the previous behavior, user must manually set
`JUPYTERHUB_README="$JUPYTERHUB_USER_DATA_DIR/README.ipynb"` in
`env.local`.