pavics-compose.sh: warn when adir in EXTRA_CONF_DIRS do not exist
Do not exit script immediately since this most probably due to a typo in
a new conf_dir and exit immediately means the whole platform will be
broken.

If we just warn and continue, the new conf_dir will not be active but at
least all the existing will work.

It's not ideal but less bad in the context of unattended autodeploy
where no human is available to take corrective action right away.