back-compat vars: fix wrong old value overriding new value during child invocation
This bug is mostly visible with child invocation.  First level
invocation, to be affected, need to use the old var name, which means
external repos only because the current repo has all switched to use the
new var names.

Previously `set_old_backwards_compatible_variables` was called before
`read_env_local` which means old var names were initialized with default
values only, not the updated values from `env.local`.  On child
invocation, since the old var already exists, it overrides the new var
even if the new var is properly set in `env.local`.

For the autodeploy process, `triggerdeploy.sh` is the first level
invocation and `autodeploy.sh` is the child invocation.  It blows up in
the child invocation.

Most other scripts do not have a child invocation so they all work fine
and that's why this back-compat bug is discovered so late.