autodeploy: fix unable to git pull newer env.local
This assume the env.local file is part of
BIRDHOUSE_AUTODEPLOY_EXTRA_REPOS.

Previously, the env.local file could be a symlink to anywhere.  Now it
can only be a symlink to a file under one of the folders listed in
BIRDHOUSE_AUTODEPLOY_EXTRA_REPOS.

But the newer docker image seems to not allow both volume-mounting the
parent dir and the file at the same time.  Can only have one of the two.